BM1370 hardware frequency table, 2D voltage exploration, 10 mV steps
Frequency navigation:
- Added BM1370_FREQ_TABLE (250 entries, 400–1100 MHz) generated from the
  chip's PLL formula: 25*N/D for D ∈ {1,2,3,4,6,8}, where D=4/8 are only
  valid ≤737.5 MHz — matching the hardware's actual stepping granularity
  (~1–4 MHz below 737.5, ~4.17 MHz above)
- Added _freq_snap / _freq_next / _freq_prev helpers for table navigation
- All frequency arithmetic replaced with table lookups so the tuner only
  ever targets frequencies the ASIC's PLL can actually produce
- Fast mode jumps 6 entries (~12.5 MHz) for initial discovery; slow mode
  steps 1 entry at a time for precision tuning near limits
- miner_client.apply_settings() now accepts float frequency

Voltage exploration:
- STEP_VOLT changed from fast=50/slow=25 to fast=20/slow=10 mV
- Removed voltage_bumped boolean — voltage now bumps as many times as
  needed until cfg.max_voltage is reached, enabling true 2D (freq × volt)
  parameter space exploration
- After frequency ceiling at any voltage level, voltage increments and
  frequency re-explores from best_freq+step — not just a one-shot bump

+++ app/miner_client.py
@@ -27,7 +27,7 @@ class MinerClient:
             pass
         return None
 
-    async def apply_settings(self, frequency: int, voltage: int) -> bool:
+    async def apply_settings(self, frequency: float, voltage: int) -> bool:
         """PATCH new freq/voltage to the miner."""
         try:
             async with httpx.AsyncClient(timeout=TIMEOUT) as c:
